# openapi.yaml
openapi: 3.0.3
info:
  title: Procurement RAG Agent API
  description: |
    Enhanced Procurement Analysis API with intelligent hybrid RAG system supporting:
    - Dynamic vendor lookup (no hardcoding)
    - Category-based analysis and supplier recommendations
    - SQL and semantic search capabilities
    - Advanced query classification and routing
    
    **Key Features:**
    - Handles category questions like "top vendor for laptops"
    - Supplier recommendations: "what supplier should I use for chairs?"
    - Dynamic fuzzy vendor matching
    - Enhanced query classification with 7 types
    - Comprehensive caching system
  version: 2.0.0
  contact:
    name: Procurement RAG Agent
    email: support@example.com
servers:
  - url: https://watsonx.orchestrate-procurement.ngrok.app
    description: Procurement RAG API via ngrok
...
paths:
...
  /ask:
    post:
      tags: [Core]
      summary: Main Query Endpoint (Enhanced Hybrid System)
      description: |
        Primary endpoint using intelligent hybrid system with enhanced capabilities:
        - Category-based queries (e.g., "top vendor for laptops")
        - Supplier recommendations (e.g., "best supplier for office chairs")
        - Dynamic vendor matching without hardcoded names
        - Automatic SQL vs RAG routing with 7 query types
        - Built-in caching and optimization
      operationId: askQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  description: Natural language procurement question
                  example: "What is our top vendor for laptops?"
                verify:
                  type: boolean
                  description: Include verification analysis
                  default: false
            examples:
              category_query:
                summary: Category-based query
                value:
                  question: "What is our top vendor for laptops?"
              supplier_recommendation:
                summary: Supplier recommendation
                value:
                  question: "What supplier should I use for office chairs?"
              category_analysis:
                summary: Category analysis
                value:
                  question: "What category has the highest number of orders?"
              general_query:
                summary: General procurement query
                value:
                  question: "What is our total spending this year?"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /top-vendors:
    get:
      tags: [Analysis]
      summary: Get Top Vendors by Spending
      description: Direct SQL query for top vendors by total spending
      operationId: getTopVendors
      parameters:
        - name: n
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of top vendors to return
      responses:
        '200':
          description: Top vendors list
          content:
            application/json:
              schema:
                type: object
                properties:
                  top_vendors:
                    type: array
                    items:
                      $ref: '#/components/schemas/TopVendorInfo'
                  total_combined:
                    type: number
                    description: Combined spending of returned vendors
                  count:
                    type: integer

  # ============================================
  # ADMINISTRATION ENDPOINTS
  # ============================================
...
components:
  schemas:
    QueryResponse:
      type: object
      properties:
        source:
          type: string
          description: Data source used (SQL, Discovery/RAG, Hybrid)
          example: "SQL"
        query_type:
          type: string
          description: Classified query type
          enum: [aggregation, ranking, comparison, specific_lookup, category_analysis, supplier_recommendation, semantic_search]
        answer:
          type: string
          description: Main response answer
        summary:
          type: string
          description: Summary of the response
        formatted_answer:
          type: string
          description: Formatted answer for display
        records_analyzed:
          type: integer
          description: Number of records analyzed
        cache_hit:
          type: boolean
          description: Whether result came from cache
        sub_queries:
          type: integer
          description: Number of sub-queries for complex queries
        verification:
          $ref: '#/components/schemas/VerificationInfo'
...
    TopVendorInfo:
      type: object
      properties:
        vendor:
          type: string
        total_spending:
          type: number
        order_count:
          type: integer
        avg_order:
          type: number
...
    VerificationInfo:
      type: object
      properties:
        confidence_score:
          type: number
          description: Confidence score (0-1)
        checks_performed:
          type: integer
        discrepancies:
          type: integer
        error:
          type: string
          description: Error message if verification failed
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        suggestion:
          type: string
          description: Helpful suggestion (if applicable)
        help:
          type: string
          description: Usage help (if applicable)